Turn some Specflow tests from round tests into tournament tests
