+++ Slask.SpecFlow.IntegrationTests/DomainTests/Tournament.feature
@@ -1,7 +1,7 @@
-﻿Feature: Round
-	Does a bunch of general tests on rounds
+﻿Feature: Tournament
+	Does a bunch of tests on a tournament as a whole
 
-@RoundTag
+@TournamentTag
 Scenario: Cannot register new players references when tournament has begun
 	Given a tournament named "GSL 2019" with users "Stålberto" added to it
 		And created tournament 0 adds rounds
@@ -25,3 +25,15 @@ Scenario: Cannot exclude players references when tournament has begun
 			| 0                | 0           | 0           | 
 		And players "First" is excluded from round 0 
 	Then created tournament 0 should contain exactly these player references with names: "First, Second, Third, Fourth, Fifth, Sixth, Seventh, Eighth"
+
+Scenario: When first round is removed the existing player references are transfered to the new first round
+	Given a tournament named "GSL 2019" has been created
+		And created tournament 0 adds rounds
+			| Round type  | Round name          | Best of | Advancing per group count | Players per group count |
+			| Round robin | Round robin round 1 | 3       | 2                         | 4                       |
+			| Round robin | Round robin round 2 | 3       | 2                         | 4                       |
+			| Round robin | Round robin round 3 | 3       | 1                         | 2                       |
+		And players "Maru, Stork, Taeja, Rain, Bomber, FanTaSy, Stephano, Thorzain" is registered to round 0
+	When round 0 is removed from tournament 0
+	Then participating players in round 0 should be exactly "Maru, Stork, Taeja, Rain, Bomber, FanTaSy, Stephano, Thorzain"
+		And tournament 0 contains 2 rounds
